Add optional project ID to key and entity converters
Summary:
As part of DIST-807, we would like to reuse the existing conversion functions
from kotlin-datastore keys to google-cloud-datastore keys, but in such a way
that we don't have to initialize the DBEnv in order to do so. This modifies the
key conversion functions so that they can accept a `projectId` parameter.

Test Plan: - `./gradlew test`

Reviewers: lizfaubell, dhruv, csilvers, samuel

Reviewed By: lizfaubell

Subscribers: colin, benkraft

Differential Revision: https://phabricator.khanacademy.org/D58132